# SINGA-490 Optimization on GPU Malloc and Cudastream